+++ core/document.ts
@@ -1,5 +1,5 @@
 import { TransformError } from "../storage/transformers/transform_error.js";
-import type { CMSContent, Data, Entry } from "../types.ts";
+import type { CMSContent, Data, Entry, PreviewURL } from "../types.ts";
 
 export interface DocumentOptions {
   name?: string;
@@ -7,7 +7,7 @@ export interface DocumentOptions {
   description?: string;
   entry: Entry;
   fields: Lume.CMS.ResolvedField;
-  url?: string;
+  previewURL?: PreviewURL;
   views?: string[] | ((data?: Data) => string[] | undefined);
   edit?: boolean;
 }
@@ -22,7 +22,7 @@ export default class Document {
   description?: string;
   #entry: Entry;
   #fields: Lume.CMS.ResolvedField;
-  url?: string;
+  previewURL?: PreviewURL;
   views?: string[] | ((data?: Data) => string[] | undefined);
   permissions: Permissions;
 
@@ -32,7 +32,7 @@ export default class Document {
     this.description = options.description;
     this.#entry = options.entry;
     this.#fields = options.fields;
-    this.url = options.url;
+    this.previewURL = options.previewURL;
     this.views = options.views;
     this.permissions = {
       edit: options.edit ?? true,